Split main window into controller and view
