+++ zapzap/controllers/main_window/__init__.py
@@ -1,0 +1,1 @@
+from .main_window_controller import MainWindowController
