+++ zapzap/views/main_window/__init__.py
@@ -1,0 +1,1 @@
+from .main_window_view import MainWindowView
