+++ zapzap/views/main_window/main_window_view.py
@@ -1,0 +1,168 @@
+"""Main window view for ZapZap."""
+
+from gettext import gettext as _
+
+from PyQt6.QtGui import QAction
+from PyQt6.QtWidgets import QHBoxLayout
+from PyQt6.QtWidgets import QMainWindow
+from PyQt6.QtWidgets import QMenu
+from PyQt6.QtWidgets import QMenuBar
+from PyQt6.QtWidgets import QStackedWidget
+from PyQt6.QtWidgets import QWidget
+
+
+class MainWindowView(QMainWindow):
+    """Main window structure and actions without application behavior."""
+
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self._setup_view()
+        self.retranslate_ui()
+
+    def _setup_view(self):
+        self.setObjectName("MainWindow")
+        self.resize(1000, 600)
+        self.setMinimumSize(200, 200)
+
+        self.centralwidget = QWidget(self)
+        self.centralwidget.setObjectName("centralwidget")
+        self.horizontalLayout = QHBoxLayout(self.centralwidget)
+        self.horizontalLayout.setContentsMargins(0, 0, 0, 0)
+        self.horizontalLayout.setSpacing(0)
+        self.horizontalLayout.setObjectName("horizontalLayout")
+
+        self.stackedWidget = QStackedWidget(self.centralwidget)
+        self.stackedWidget.setObjectName("stackedWidget")
+        self.horizontalLayout.addWidget(self.stackedWidget)
+        self.setCentralWidget(self.centralwidget)
+
+        self.menubar = QMenuBar(self)
+        self.menubar.setObjectName("menubar")
+        self.menuFile = QMenu(self.menubar)
+        self.menuFile.setObjectName("menuFile")
+        self.menuView = QMenu(self.menubar)
+        self.menuView.setObjectName("menuView")
+        self.menuChat = QMenu(self.menubar)
+        self.menuChat.setObjectName("menuChat")
+        self.menuUsers = QMenu(self.menubar)
+        self.menuUsers.setObjectName("menuUsers")
+        self.menuHelp = QMenu(self.menubar)
+        self.menuHelp.setObjectName("menuHelp")
+        self.setMenuBar(self.menubar)
+
+        self._create_actions()
+        self._compose_menus()
+
+    def _create_actions(self):
+        self.actionQuit = QAction(self)
+        self.actionQuit.setObjectName("actionQuit")
+        self.actionHide = QAction(self)
+        self.actionHide.setObjectName("actionHide")
+        self.actionSettings = QAction(self)
+        self.actionSettings.setObjectName("actionSettings")
+        self.actionReset_zoom = QAction(self)
+        self.actionReset_zoom.setObjectName("actionReset_zoom")
+        self.actionZoom_in = QAction(self)
+        self.actionZoom_in.setObjectName("actionZoom_in")
+        self.actionZoom_out = QAction(self)
+        self.actionZoom_out.setObjectName("actionZoom_out")
+        self.actionToggle_full_screen = QAction(self)
+        self.actionToggle_full_screen.setObjectName("actionToggle_full_screen")
+        self.actionOpen_DevTools = QAction(self)
+        self.actionOpen_DevTools.setObjectName("actionOpen_DevTools")
+        self.actionGrid_view = QAction(self)
+        self.actionGrid_view.setObjectName("actionGrid_view")
+        self.actionToggle_sidebar = QAction(self)
+        self.actionToggle_sidebar.setCheckable(True)
+        self.actionToggle_sidebar.setObjectName("actionToggle_sidebar")
+        self.actionTheme_auto = QAction(self)
+        self.actionTheme_auto.setCheckable(True)
+        self.actionTheme_auto.setObjectName("actionTheme_auto")
+        self.actionTheme_light = QAction(self)
+        self.actionTheme_light.setCheckable(True)
+        self.actionTheme_light.setObjectName("actionTheme_light")
+        self.actionTheme_dark = QAction(self)
+        self.actionTheme_dark.setCheckable(True)
+        self.actionTheme_dark.setObjectName("actionTheme_dark")
+        self.actionReload = QAction(self)
+        self.actionReload.setObjectName("actionReload")
+        self.actionNew_chat = QAction(self)
+        self.actionNew_chat.setObjectName("actionNew_chat")
+        self.actionBy_phone_number = QAction(self)
+        self.actionBy_phone_number.setObjectName("actionBy_phone_number")
+        self.actionShortcuts = QAction(self)
+        self.actionShortcuts.setObjectName("actionShortcuts")
+        self.actionSobre_o_ZapZap = QAction(self)
+        self.actionSobre_o_ZapZap.setObjectName("actionSobre_o_ZapZap")
+
+    def _compose_menus(self):
+        self.menuFile.addAction(self.actionSettings)
+        self.menuFile.addSeparator()
+        self.menuFile.addAction(self.actionHide)
+        self.menuFile.addAction(self.actionQuit)
+        self.menuView.addAction(self.actionGrid_view)
+        self.menuView.addAction(self.actionOpen_DevTools)
+        self.menuView.addSeparator()
+        self.menuView.addAction(self.actionToggle_sidebar)
+        self.menuView.addAction(self.actionToggle_full_screen)
+        self.menuView.addSeparator()
+        self.menuView.addAction(self.actionTheme_auto)
+        self.menuView.addAction(self.actionTheme_light)
+        self.menuView.addAction(self.actionTheme_dark)
+        self.menuView.addSeparator()
+        self.menuView.addAction(self.actionReset_zoom)
+        self.menuView.addAction(self.actionZoom_in)
+        self.menuView.addAction(self.actionZoom_out)
+        self.menuChat.addAction(self.actionReload)
+        self.menuChat.addSeparator()
+        self.menuChat.addAction(self.actionNew_chat)
+        self.menuChat.addAction(self.actionBy_phone_number)
+        self.menuHelp.addAction(self.actionShortcuts)
+        self.menuHelp.addAction(self.actionSobre_o_ZapZap)
+        self.menubar.addAction(self.menuFile.menuAction())
+        self.menubar.addAction(self.menuView.menuAction())
+        self.menubar.addAction(self.menuChat.menuAction())
+        self.menubar.addAction(self.menuUsers.menuAction())
+        self.menubar.addAction(self.menuHelp.menuAction())
+
+    def retranslate_ui(self):
+        self.setWindowTitle(_("ZapZap"))
+        self.menuFile.setTitle(_("File"))
+        self.menuView.setTitle(_("View"))
+        self.menuChat.setTitle(_("Chat"))
+        self.menuUsers.setTitle(_("Users"))
+        self.menuHelp.setTitle(_("Help"))
+        self.actionQuit.setText(_("Quit"))
+        self.actionQuit.setShortcut(_("Ctrl+Q"))
+        self.actionHide.setText(_("Hide"))
+        self.actionHide.setShortcut(_("Ctrl+W"))
+        self.actionSettings.setText(_("Settings"))
+        self.actionSettings.setShortcut(_("Ctrl+P"))
+        self.actionReset_zoom.setText(_("Reset zoom"))
+        self.actionReset_zoom.setShortcut(_("Ctrl+0"))
+        self.actionZoom_in.setText(_("Zoom in"))
+        self.actionZoom_in.setShortcut(_("Ctrl++"))
+        self.actionZoom_out.setText(_("Zoom out"))
+        self.actionZoom_out.setShortcut(_("Ctrl+-"))
+        self.actionToggle_full_screen.setText(_("Toggle full screen"))
+        self.actionToggle_full_screen.setShortcut(_("F11"))
+        self.actionOpen_DevTools.setText(_("Open DevTools"))
+        self.actionOpen_DevTools.setShortcut(_("Ctrl+Shift+I"))
+        self.actionGrid_view.setText(_("Grid View"))
+        self.actionGrid_view.setShortcut(_("Ctrl+G"))
+        self.actionToggle_sidebar.setText(_("Show sidebar"))
+        self.actionTheme_auto.setText(_("Automatic theme"))
+        self.actionTheme_auto.setShortcut(_("Ctrl+Alt+1"))
+        self.actionTheme_light.setText(_("Light theme"))
+        self.actionTheme_light.setShortcut(_("Ctrl+Alt+2"))
+        self.actionTheme_dark.setText(_("Dark theme"))
+        self.actionTheme_dark.setShortcut(_("Ctrl+Alt+3"))
+        self.actionReload.setText(_("Reload"))
+        self.actionReload.setShortcut(_("F5"))
+        self.actionNew_chat.setText(_("New chat"))
+        self.actionNew_chat.setShortcut(_("Ctrl+N"))
+        self.actionBy_phone_number.setText(_("By number phone"))
+        self.actionBy_phone_number.setShortcut(_("Ctrl+M"))
+        self.actionShortcuts.setText(_("Keyboard shortcuts"))
+        self.actionShortcuts.setShortcut(_("Ctrl+Shift+/, Ctrl+?"))
+        self.actionSobre_o_ZapZap.setText(_("About"))
